// types.ts
export enum Screen {
  DASHBOARD = 'DASHBOARD',
  CLIENTS_ROOT = 'CLIENTS_ROOT',
  ADD_CLIENT = 'ADD_CLIENT',
  CLIENT_DETAILS = 'CLIENT_DETAILS',
  ADD_PRODUCT_REQUEST = 'ADD_PRODUCT_REQUEST',
  ADD_ADDRESS = 'ADD_ADDRESS',
  ORDERS = 'ORDERS',
  HOME = 'HOME',
  FEED = 'FEED',
  LEADERBOARD = 'LEADERBOARD',
  ME = 'ME',
  DEAL_DETAIL = 'DEAL_DETAIL',
  JOIN_IN_DETAIL = 'JOIN_IN_DETAIL'
}